rename package name to openvswitch
