+++ openvswitch/ovs_test.go
@@ -1,4 +1,4 @@
-package ovs
+package openvswitch
 
 import (
 	"os"
